ENH: Use the docs to expend interfaces in sphcom.f
Several of the functions have documentation I can use to expand the interfaces
Some of them are wrappers around another function I can use for the same purpose

Specifying ranks required some changes in other functions.
One function took a 1-d input and passed part of it to a 2-d argument
Fixing that required a pointer to reinterpret the data.